feat: logger

+++ backend/migrations/20230930053536_UsersCreationMigration.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-                       id int not null primary key,
+                       id bigserial not null primary key,
                        email varchar not null unique,
                        password varchar not null
 );
